feat: lenient matching
